Updated README.md instructions and fixed an error.
Previous attempts to install with pacman would be met with the error "invalid or corrupted package (checksum)", despite downloading the correct .db file. The server URL was changed to https://raw.githubusercontent.com as the base after using the wrong raw URL. As a result, the checksum is valid upon installation and the programs run as expected.